Update e2e-tests workflow to build binaries
The e2e-tests workflow has been updated to include a job for building binaries. This is a crucial step to run end-to-end tests as it ensures the latest version of the application is being tested. This job only runs on non-dependabot branches and runs on ubuntu with the latest version of teleport-buildbox. The binaries are then uploaded as an artifact for use in subsequent jobs. This change improves the thoroughness of our tests and ensures they are running against the most up-to-date version of the codebase.