Add retention days and chmod to binaries in e2e-tests workflow
The changes in this commit updated the workflow for end-to-end testing. A retention days field was added to keep the generated build files for 1 day only. Also, the 'chmod +x' commands were added to ensure the binaries 'teleport', 'tctl', and 'tsh' have execute permissions. This is necessary to avoid potential permission issues during execution of the end-to-end tests.